Version 1.3.1 (#57)
* feat: replace Bonjour service with udp-discovery for service discovery

* feat: update README and server configuration for environment variables

* feat: initialize Discovery with configurable port from environment variables

* Fixed an issue with horizontal scroll

Added overflow-x hidden to the body to prevent out of bounds elements to cause horizontal scrolling
Added a scrollbar-width: thin on every scrollbar purely for aesthetic reasons.

* feat: update Dockerfile and README to expose UDP port 44201 for service discovery

* Added an option to remove the libraries section from the home screen

An option to remove the libraries section from the home screen as it is redundant with the top nav. I made a new UI section in the settings page, it may or may not be the most organized idea, let me know!

* Added an option to remove the libraries section from the home screen

An option to remove the libraries section from the home screen as it is redundant with the top nav. I made a new UI section in the settings page, it may or may not be the most organized idea, let me know!

* Moved "Disable Home Libraries Section" to Libraries Settings

* Revert "Moved "Disable Home Libraries Section" to Libraries Settings"

This reverts commit d2b26be75aa43149931c2619a467c69ccb3d7553.

* Revert "Added an option to remove the libraries section from the home screen"

This reverts commit f5748919fff9b20dd63c36d658f6cc49fcff60bd.

* Moved "Disable Home Libraries Section" to Libraries Settings

* Add platform detection functionality and cache implementation

* Enhance getXPlexProps function to use platform detection

* Add device name retrieval and update platform cache structure

* Use HLS instead of DASHJS for Desktop

* Fixed the videos in the hero banner not taking the whole height/width

Added an objectFit: cover to the HeroDisplay Component so that the videos takes the whole space, also moved the z-index property as it was no being applied when not directly on the video component.

* Fixed hero video controls buttons being displayed on top of the app bar

Changed the z-index of the buttons of the hero to not make them appear over the app bar

---------

Co-authored-by: PixyTM <43484698+PixyTM@users.noreply.github.com>